Move platformPrefixIcon into its own dedicated module
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
