+++ modules/icon/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export {Icon} from 'react-native-vector-icons/Ionicons'
+
+export {platformPrefixIconName} from './platform-prefix-icon-name'
