+++ modules/icon/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@frogpond/icon",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "author": "",
+  "license": "ISC",
+  "scripts": {
+    "test": "jest"
+  },
+  "peerDependencies": {
+    "react-native": "^0.57.0",
+    "react-native-vector-icons": "^6.0.0"
+  }
+}
