+++ modules/icon/platform-prefix-icon-name.js
@@ -1,0 +1,16 @@
+// @flow
+
+import {Platform} from 'react-native'
+import {Icon} from '.'
+
+export function platformPrefixIconName(name: string) {
+	let isAvailable = Icon.hasIcon(name)
+	let isAvailableOnBothPlatforms =
+		Icon.hasIcon(`ios-${name}`) && Icon.hasIcon(`md-${name}`)
+
+	if (isAvailable && !isAvailableOnBothPlatforms) {
+		return name
+	}
+
+	return Platform.OS === 'ios' ? `ios-${name}` : `md-${name}`
+}
